fix: improve WebSocket health detection and connection handling
- track the timestamp of the last received WebSocket data
- reduce the keepalive interval to improve dead-connection detection
- report closed sockets, failed keepalives, and send failures consistently
- stop the keepalive timer when closing the connection
- adapt WebSocket event handlers to PureWebSockets 4
- update BouncyCastle, Newtonsoft.Json, PureWebSockets, and crypto dependencies
- widen miniserver user rights from int to long
- remove unused imports and apply consistent code formatting

+++ Data/LoxApp/MiniserverInfo.cs
@@ -1,7 +1,4 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Loxone.Api.Data.LoxApp
 {
@@ -11,7 +8,7 @@ namespace Loxone.Api.Data.LoxApp
         public string Uuid { get; set; }
         public bool isAdmin { get; set; }
         public bool changePassword { get; set; }
-        public int UserRights { get; set; }
+        public long UserRights { get; set; }
     }
     public class MiniserverInfo
     {
